content: generateTrackChronologyLinks
Just extracts logic from generateTrackInfoPage, for now.
