+++ src/content/dependencies/generateTrackInfoPage.js
@@ -1,8 +1,5 @@
-import {sortAlbumsTracksChronologically, sortFlashesChronologically}
-  from '#sort';
+import {sortFlashesChronologically} from '#sort';
 import {empty, stitchArrays} from '#sugar';
-
-import getChronologyRelations from '../util/getChronologyRelations.js';
 
 export default {
   contentDependencies: [
@@ -12,7 +9,6 @@ export default {
     'generateAlbumSecondaryNav',
     'generateAlbumSidebar',
     'generateAlbumStyleRules',
-    'generateChronologyLinks',
     'generateColorStyleAttribute',
     'generateCommentarySection',
     'generateContentHeading',
@@ -20,13 +16,13 @@ export default {
     'generatePageLayout',
     'generateRelativeDatetimestamp',
     'generateTrackAdditionalNamesBox',
+    'generateTrackChronologyLinks',
     'generateTrackCoverArtwork',
     'generateTrackList',
     'generateTrackListDividedByGroups',
     'generateTrackReleaseInfo',
     'generateTrackSocialEmbed',
     'linkAlbum',
-    'linkArtist',
     'linkFlash',
     'linkTrack',
     'transformContent',
@@ -55,51 +51,6 @@ export default {
     relations.socialEmbed =
       relation('generateTrackSocialEmbed', track);
 
-    relations.artistChronologyContributions =
-      getChronologyRelations(track, {
-        contributions: [
-          ...track.artistContribs ?? [],
-          ...track.contributorContribs ?? [],
-        ],
-
-        linkArtist: artist => relation('linkArtist', artist),
-        linkThing: track => relation('linkTrack', track),
-
-        getThings(artist) {
-          const getDate = thing => thing.date;
-
-          const things = [
-            ...artist.tracksAsArtist,
-            ...artist.tracksAsContributor,
-          ].filter(getDate);
-
-          return sortAlbumsTracksChronologically(things, {getDate});
-        },
-      });
-
-    relations.coverArtistChronologyContributions =
-      getChronologyRelations(track, {
-        contributions: track.coverArtistContribs ?? [],
-
-        linkArtist: artist => relation('linkArtist', artist),
-
-        linkThing: trackOrAlbum =>
-          (trackOrAlbum.album
-            ? relation('linkTrack', trackOrAlbum)
-            : relation('linkAlbum', trackOrAlbum)),
-
-        getThings(artist) {
-          const getDate = thing => thing.coverArtDate ?? thing.date;
-
-          const things = [
-            ...artist.albumsAsCoverArtist,
-            ...artist.tracksAsCoverArtist,
-          ].filter(getDate);
-
-          return sortAlbumsTracksChronologically(things, {getDate});
-        },
-      }),
-
     relations.albumLink =
       relation('linkAlbum', track.album);
 
@@ -109,8 +60,8 @@ export default {
     relations.albumNavAccent =
       relation('generateAlbumNavAccent', track.album, track);
 
-    relations.chronologyLinks =
-      relation('generateChronologyLinks');
+    relations.trackChronologyLinks =
+      relation('generateTrackChronologyLinks', track);
 
     relations.secondaryNav =
       relation('generateAlbumSecondaryNav', track.album);
@@ -580,18 +531,7 @@ export default {
           }),
 
         navContent:
-          relations.chronologyLinks.slots({
-            chronologyInfoSets: [
-              {
-                headingString: 'misc.chronology.heading.track',
-                contributions: relations.artistChronologyContributions,
-              },
-              {
-                headingString: 'misc.chronology.heading.coverArt',
-                contributions: relations.coverArtistChronologyContributions,
-              },
-            ],
-          }),
+          relations.trackChronologyLinks,
 
         secondaryNav:
           relations.secondaryNav
